+++ docs/history.md
@@ -3,6 +3,7 @@
 ### *v0.78.16* - March 12th, 2022
 
 - Fixed the bug where in seeded races, you would gain coins upon reaching a new floor if you had Deep Pockets and more than 100 coins. (Thanks Gamonymous)
+- Fixed the bug where extra starting items for the first floor were not properly being removed in situations where the Treasure Room was banned. (Thanks Gamonymous)
 
 ### *v0.78.15* - March 12th, 2022